AIPrioritization/requirements.txt
fastapi==0.104.1
redis==4.6.0
uvicorn[standard]==0.24.0
pydantic-settings==2.1.0
pydantic==2.5.0
openai>=1.0.0
python-multipart==0.0.6
requests==2.31.0
aiohttp==3.9.1
numpy==1.26.2
scikit-learn==1.3.2
nats-py==2.6.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3

# Monitoring and Observability (Optional)
# prometheus-client==0.19.0
# opentelemetry-api==1.21.0
# opentelemetry-sdk==1.21.0
